refactor: fetch_page_metadata를 enrichment_network로 분리 + 6/2 split 완성 (P2-A batch 3)
docs/refactoring-plan-2026-07.md §1.4 3단계에 따라 fetch_page_metadata 정의를
파사드(common/enrichment.py)에서 common/enrichment_network.py로 이동(byte-identical).
fetch_page_description·_fetch_and_parse_page는 파사드 유지(Correction A), 재-export
바인딩으로 위임. __all__ 불변(12).

fetch_page_metadata가 network로 이동하면서 내부 헬퍼 호출이 전부 network
네임스페이스로 해석 → 이를 구동하는 @patch을 같은 커밋에서 network로 재배치:
- requests.get 29 → network (Correction D; facade의 유일 requests 사용처였음)
- _resolve_google_news_url 2(:1263,:1981) → network = 6/2 split 완성
  (나머지 6은 fetch_images/descriptions_concurrent 경유라 facade 유지)
- sanitize_mojibake 3 → network (Correction B)
- is_private_url 6 → network (순수 함수 false-green 리스크; 리뷰어가 mock.called로
  facade inert / network 유효 실증), 추출기 _extract_via_* 6 → network
- fetch_page_metadata patch 16 → facade 유지(호출자 enrich_item/오케스트레이터 잔류)

부수: facade의 orphan된 import requests + encoding_guard(sanitize_mojibake·
force_utf8_if_mislabelled) 제거(ruff F401 autofix). facade requests 참조 0건 →
batch 4의 requests 제거는 조기 충족.

게이트: import OK(무순환), 715 passed, ruff clean, basedpyright 0 errors.
독립 코드리뷰 APPROVE(byte-identical 이동 + REPL 네임스페이스 실증).
