+++ tests/test_enrichment.py
@@ -399,7 +399,7 @@ class TestFetchPageMetadata:
         for key in ("description", "image", "published_time", "author", "section"):
             assert key in result
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_published_time_extracted(self, mock_get):
         html = """<html><head>
         <meta property="article:published_time" content="2026-04-21T09:30:00+09:00" />
@@ -412,7 +412,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["published_time"] == "2026-04-21T09:30:00+09:00"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_author_extracted(self, mock_get):
         html = """<html><head>
         <meta property="article:author" content="홍길동 기자" />
@@ -425,7 +425,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["author"] == "홍길동 기자"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_section_extracted(self, mock_get):
         html = """<html><head>
         <meta property="article:section" content="경제" />
@@ -438,7 +438,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["section"] == "경제"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_modified_time_fallback_when_published_absent(self, mock_get):
         html = """<html><head>
         <meta property="article:modified_time" content="2026-04-21T10:00:00Z" />
@@ -451,7 +451,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["published_time"] == "2026-04-21T10:00:00Z"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_http_error_returns_empty(self, mock_get):
         import requests as req_mod
 
@@ -459,7 +459,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["description"] == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_og_description_extracted(self, mock_get):
         html = """<html><head>
         <meta property="og:description" content="Bitcoin surged 10% on institutional demand today." />
@@ -472,7 +472,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert "Bitcoin" in result["description"]
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_meta_description_extracted(self, mock_get):
         html = """<html><head>
         <meta name="description" content="Ethereum network upgrade brings major improvements to the ecosystem." />
@@ -484,8 +484,8 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert "Ethereum" in result["description"]
 
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
     def test_irrelevant_meta_description_rejected_with_title(self, mock_get, _mock_private):  # noqa: PT019
         html = """<html><head>
         <meta name="description" content="Apple unveils new iPhone lineup at annual event." />
@@ -497,7 +497,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article", title="Bitcoin ETF approval drives market rally")
         assert result["description"] == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_og_image_extracted(self, mock_get):
         html = """<html><head>
         <meta property="og:image" content="https://example.com/article-image.jpg" />
@@ -510,7 +510,7 @@ class TestFetchPageMetadata:
         result = fetch_page_metadata("https://example.com/article")
         assert result["image"] == "https://example.com/article-image.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_short_description_skipped(self, mock_get):
         html = """<html><head>
         <meta name="description" content="Short." />
@@ -523,7 +523,7 @@ class TestFetchPageMetadata:
         # Short meta desc < 20 chars is skipped; body paragraph may be extracted
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_http_status_error_returns_empty(self, mock_get):
         import requests as req_mod
 
@@ -1260,7 +1260,7 @@ class TestFetchDescriptionsConcurrentExtended:
 class TestFetchPageMetadataExtended:
     """Extended tests for fetch_page_metadata covering more branches."""
 
-    @patch("common.enrichment._resolve_google_news_url")
+    @patch("common.enrichment_network._resolve_google_news_url")
     def test_google_news_url_resolved(self, mock_resolve):
         """Google News URL should be resolved before fetching."""
         from common.enrichment import fetch_page_metadata
@@ -1276,7 +1276,7 @@ class TestFetchPageMetadataExtended:
             "section": "",
         }
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_twitter_description_extracted(self, mock_get):
         """twitter:description meta tag should be used as fallback."""
         from common.enrichment import fetch_page_metadata
@@ -1291,7 +1291,7 @@ class TestFetchPageMetadataExtended:
         result = fetch_page_metadata("https://example.com/article")
         assert "Ethereum" in result["description"]
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_body_paragraph_fallback(self, mock_get):
         """If no meta description, fall back to first long body <p> tag."""
         from common.enrichment import fetch_page_metadata
@@ -1307,7 +1307,7 @@ class TestFetchPageMetadataExtended:
         result = fetch_page_metadata("https://example.com/article")
         assert "long enough paragraph" in result["description"]
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_tag_used_for_paragraphs(self, mock_get):
         """Prefers <article> tag body over generic <p> tags."""
         from common.enrichment import fetch_page_metadata
@@ -1325,7 +1325,7 @@ class TestFetchPageMetadataExtended:
         result = fetch_page_metadata("https://example.com/article")
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_twitter_image_extracted(self, mock_get):
         """twitter:image should be extracted when og:image is absent."""
         from common.enrichment import fetch_page_metadata
@@ -1977,9 +1977,9 @@ class TestNeedsEnrichmentDescEqualsTitle:
 class TestFetchPageMetadataGoogleNewsResolved:
     """Cover line 376: url = resolved after Google News resolution."""
 
-    @patch("common.enrichment.requests.get")
-    @patch("common.enrichment._resolve_google_news_url")
-    @patch("common.enrichment.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
+    @patch("common.enrichment_network._resolve_google_news_url")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
     def test_google_news_resolved_url_used_for_fetch(self, mock_private, mock_resolve, mock_get):
         """Line 376: resolved URL replaces google news URL for metadata fetch."""
         mock_resolve.return_value = "https://real-article.com/news/story"
@@ -2008,7 +2008,7 @@ class TestFetchPageMetadataGoogleNewsResolved:
 class TestFetchPageMetadataReadabilityAndArticleBody:
     """Cover readability paragraph extraction (line 428) and BS4 article body (433-485)."""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_readability_used_when_available_five_paragraphs(self, mock_get):
         """Line 428: readability break after 5 paragraphs."""
         # Build HTML with 7 long paragraphs
@@ -2035,7 +2035,7 @@ class TestFetchPageMetadataReadabilityAndArticleBody:
             # Should have truncated after 5 paragraphs
             assert len(result["description"]) > 0
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_tag_body_extraction(self, mock_get):
         """Lines 461-475: <article> tag paragraph extraction."""
         html = """<html><head></head><body>
@@ -2054,7 +2054,7 @@ class TestFetchPageMetadataReadabilityAndArticleBody:
             result = fetch_page_metadata("https://example.com/article")
         assert "financial markets" in result["description"] or isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_class_body_extraction(self, mock_get):
         """Lines 448-455: article-body class used when no <article> tag."""
         html = """<html><head></head><body>
@@ -2071,7 +2071,7 @@ class TestFetchPageMetadataReadabilityAndArticleBody:
             result = fetch_page_metadata("https://example.com/article")
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_fallback_p_tags_when_no_article(self, mock_get):
         """Lines 478-485: fallback <p> outside article containers."""
         html = """<html><head></head><body>
@@ -2088,7 +2088,7 @@ class TestFetchPageMetadataReadabilityAndArticleBody:
             result = fetch_page_metadata("https://example.com/article")
         assert "standalone paragraph" in result["description"] or len(result["description"]) > 0
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_readability_exception_falls_through_to_bs4(self, mock_get):
         """Lines 435-436: readability raises non-ImportError, falls through to BS4."""
         html = """<html><head></head><body>
@@ -2315,7 +2315,7 @@ class TestGenerateSyntheticDescriptionFallbackBranches:
 class TestFetchPageMetadataRemainingLines:
     """Cover lines 434, 458-460, 464, 471, 481."""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_readability_import_error_pass_line_434(self, mock_get):
         """Line 434: readability ImportError -> pass, fall through to BS4.
 
@@ -2349,7 +2349,7 @@ class TestFetchPageMetadataRemainingLines:
 
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_exact_class_match_fallback_lines_458_460(self, mock_get):
         """Lines 458-460: exact class regex (^article|post|entry|content$) used
         when precise class pattern doesn't match."""
@@ -2367,7 +2367,7 @@ class TestFetchPageMetadataRemainingLines:
             result = fetch_page_metadata("https://example.com/article")
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_noise_element_decomposed_line_464(self, mock_get):
         """Line 464: noise.decompose() called when article has excluded class child."""
         html = """<html><head></head><body>
@@ -2385,7 +2385,7 @@ class TestFetchPageMetadataRemainingLines:
             result = fetch_page_metadata("https://example.com/article")
         assert isinstance(result["description"], str)
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_five_paragraphs_break_line_471(self, mock_get):
         """Line 471: break after collecting 5 paragraphs from article body."""
         # Build 7 paragraphs inside <article> so the break at 5 is triggered
@@ -2403,7 +2403,7 @@ class TestFetchPageMetadataRemainingLines:
             result = fetch_page_metadata("https://example.com/article")
         assert len(result["description"]) > 0
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_fallback_p_skips_excluded_parent_line_481(self, mock_get):
         """Line 481: continue when fallback <p> has excluded-class parent."""
         html = """<html><head></head><body>
@@ -3232,7 +3232,7 @@ class TestExtractViaBs4ArticleEmptyParagraphs:
 # ---------------------------------------------------------------------------
 class TestFetchPageMetadataSSRFBlock:
     def test_private_url_returns_empty_dict(self):
-        with patch("common.enrichment.is_private_url", return_value=True):
+        with patch("common.enrichment_network.is_private_url", return_value=True):
             result = fetch_page_metadata("https://internal.corp/article")
         assert result == {
             "description": "",
@@ -3247,8 +3247,8 @@ class TestFetchPageMetadataSSRFBlock:
 # Lines 864, 873, 882: readability/bs4/paragraph unrelated-to-title rejection
 # ---------------------------------------------------------------------------
 class TestFetchPageMetadataUnrelatedDescRejected:
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
     def test_readability_desc_unrelated_to_title_rejected(self, mock_get, _mock_priv):  # noqa: PT019
         # Produce a page where og: desc is empty but readability succeeds with unrelated text
         html = (
@@ -3268,10 +3268,10 @@ class TestFetchPageMetadataUnrelatedDescRejected:
         # readability is available; patch _extract_via_readability to return an unrelated desc
         with (
             patch(
-                "common.enrichment._extract_via_readability",
+                "common.enrichment_network._extract_via_readability",
                 return_value="Apple released a new MacBook with an M4 chip for creative professionals.",
             ),
-            patch("common.enrichment.sanitize_mojibake", side_effect=lambda x: x),
+            patch("common.enrichment_network.sanitize_mojibake", side_effect=lambda x: x),
         ):
             result = fetch_page_metadata(
                 "https://example.com/article",
@@ -3280,8 +3280,8 @@ class TestFetchPageMetadataUnrelatedDescRejected:
         # Unrelated readability desc should be rejected, result description stays ""
         assert result["description"] == ""
 
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
     def test_bs4_article_desc_unrelated_to_title_rejected(self, mock_get, _mock_priv):  # noqa: PT019
         html = (
             "<html><body>"
@@ -3295,12 +3295,12 @@ class TestFetchPageMetadataUnrelatedDescRejected:
         mock_get.return_value = mock_resp
 
         with (
-            patch("common.enrichment._extract_via_readability", return_value=""),
+            patch("common.enrichment_network._extract_via_readability", return_value=""),
             patch(
-                "common.enrichment._extract_via_bs4_article",
+                "common.enrichment_network._extract_via_bs4_article",
                 return_value="Apple released a new MacBook with M4 chip for creative professionals.",
             ),
-            patch("common.enrichment.sanitize_mojibake", side_effect=lambda x: x),
+            patch("common.enrichment_network.sanitize_mojibake", side_effect=lambda x: x),
         ):
             result = fetch_page_metadata(
                 "https://example.com/article",
@@ -3308,8 +3308,8 @@ class TestFetchPageMetadataUnrelatedDescRejected:
             )
         assert result["description"] == ""
 
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
     def test_paragraph_desc_unrelated_to_title_rejected(self, mock_get, _mock_priv):  # noqa: PT019
         html = (
             "<html><body><p>Apple released a new MacBook with M4 chip targeting professional users.</p></body></html>"
@@ -3320,13 +3320,13 @@ class TestFetchPageMetadataUnrelatedDescRejected:
         mock_get.return_value = mock_resp
 
         with (
-            patch("common.enrichment._extract_via_readability", return_value=""),
-            patch("common.enrichment._extract_via_bs4_article", return_value=""),
+            patch("common.enrichment_network._extract_via_readability", return_value=""),
+            patch("common.enrichment_network._extract_via_bs4_article", return_value=""),
             patch(
-                "common.enrichment._extract_via_paragraphs",
+                "common.enrichment_network._extract_via_paragraphs",
                 return_value="Apple released a new MacBook with M4 chip for creative professionals.",
             ),
-            patch("common.enrichment.sanitize_mojibake", side_effect=lambda x: x),
+            patch("common.enrichment_network.sanitize_mojibake", side_effect=lambda x: x),
         ):
             result = fetch_page_metadata(
                 "https://example.com/article",
